Updated to the latest version of roc
