+++ README.md
@@ -1,5 +1,7 @@
-# roc-plugin-browsersync
-[![Build Status](https://travis-ci.org/rocjs/roc-plugin-browsersync.svg?branch=master)](https://travis-ci.org/rocjs/roc-plugin-browsersync)
+# roc-plugin-browsersync [![Build Status](https://travis-ci.org/rocjs/roc-plugin-browsersync.svg?branch=master)](https://travis-ci.org/rocjs/roc-plugin-browsersync)
 
 __Plugin providing Browsersync__  
-- [roc-plugin-browsersync](/packages/roc-plugin-browsersync)
+- [roc-plugin-browsersync](/extensions/roc-plugin-browsersync)
+
+---
+To be used together with [Roc](https://github.com/rocjs/roc).
