+++ extensions/roc-plugin-browsersync/src/index.js
@@ -1,0 +1,1 @@
+export roc from './roc';
