--- packages/roc-plugin-browsersync/README.md
@@ -1,8 +1,0 @@
-# roc-plugin-browsersync
-Plugin providing Browsersync.
-
-## Documentation
-- [Actions](/packages/roc-plugin-browsersync/docs/Actions.md)
-- [Commands](/packages/roc-plugin-browsersync/docs/Commands.md)
-- [Hooks](/packages/roc-plugin-browsersync/docs/Hooks.md)
-- [Settings](/packages/roc-plugin-browsersync/docs/Settings.md)
