--- packages/roc-plugin-browsersync/docs/Hooks.md
@@ -1,1 +1,0 @@
-No hooks available.
