--- packages/roc-plugin-browsersync/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "roc-plugin-browsersync",
-  "version": "1.0.0-alpha.2",
-  "description": "Plugin providing Browsersync for Roc",
-  "main": "lib/index.js",
-  "scripts": {
-    "lint": "eslint .",
-    "test": "npm run lint"
-  },
-  "files": [
-    "lib"
-  ],
-  "keywords": [
-    "roc",
-    "roc-plugin",
-    "browsersync"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/rocjs/roc-plugin-browsersync.git"
-  },
-  "author": "VG",
-  "license": "MIT",
-  "dependencies": {
-    "browser-sync": "~2.9.11",
-    "roc": "^1.0.0-rc"
-  },
-  "devDependencies": {
-    "babel-eslint": "~5.0.0",
-    "eslint": "~1.10.3",
-    "eslint-config-vgno": "~5.0.0"
-  }
-}
