--- packages/roc-plugin-browsersync/src/index.js
@@ -1,3 +1,0 @@
-export roc from './roc';
-
-export initBrowsersync from './browsersync';
